qcacld-3.0: Flush the idle shutdown before the dsc op start
Currently idle shutdown is flushed after the suspend dsc op
start. This can lead to the deadlock when the idle shutdown
work is running as it waits for the dsc ops to complete.

To mitigate this issue, flush the idle shutdown work before
starting the dsc ops in suspend sequence.

CRs-Fixed: 2941314
Change-Id: I023b9eb3543d83a30e0587cec3977e08ac7ab752
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: bagaskara815 <bagaskara815@gmail.com>